Microservicios/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyJWT==2.3.0
python-dotenv==0.19.2
python-http-client==3.3.4
pytz==2021.3
requests==2.26.0
sendgrid==6.9.2
starkbank-ecdsa==2.0.3
twilio==7.3.2
urllib3==1.26.7
Werkzeug==2.0.2
